requirements.txt
Flask==3.1.3
flask_login==0.6.3
huggingface_hub==1.8.0
python-dotenv==1.2.2
Requests==2.33.1
SQLAlchemy==2.0.48
Werkzeug==3.1.7
